+++ packages/react-docs/pages/lab/date-pickers/index.mdx
@@ -1,8 +1,6 @@
-# Date and Time Pickers
+# Date Pickers
 
-Date and Time pickers allow the user to select a date and/or time from a pre-determined set of values.
-
-## Pickers
+Date pickers allow the user to select dates from a pre-determined set of values.
 
 ```jsx noInline
 // DateInput
