+++ packages/react-lab/src/date-pickers/validation.js
@@ -1,7 +1,7 @@
 import isAfter from 'date-fns/isAfter';
 import isBefore from 'date-fns/isBefore';
 import isValid from 'date-fns/isValid';
-import isNullOrUndefined from '../../utils/isNullOrUndefined';
+import isNullOrUndefined from '../utils/isNullOrUndefined';
 
 const validateDate = (value, props) => {
   const { maxDate, minDate } = { ...props };
